+++ .gitignore
@@ -23,3 +23,5 @@ Thumbs.db
 UserInterfaceState.xcuserstate
 .env
 *.tgz
+
+docs/
