+++ demo/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <title>Meridian Web SDK</title>
+    <title>Meridian Web SDK Examples</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="robots" content="noindex, nofollow" />
     <style>
@@ -52,13 +52,9 @@
   </head>
 
   <body>
-    <h1>Meridian Web SDK</h1>
+    <h1>Meridian Web SDK Examples</h1>
     <h2>Documentation</h2>
-    <a href="https://docs.meridianapps.com/article/779-the-meridian-web-sdk"
-      >Read our documentation</a
-    >
-    <h2>Download</h2>
-    <a href="meridian-sdk.js" download>Download meridian-sdk.js</a>
+    <a href="/">Read our documentation</a>
     <h2>Examples</h2>
     <a href="basic">Basic</a>
     <a href="alternate">API as an option</a>
