+++ package.json
@@ -3,13 +3,14 @@
   "private": false,
   "version": "0.7.0",
   "description": "Web SDK for showing Aruba Meridian maps, tags, and more",
-  "main": "./dist/index.js",
-  "types": "./dist/src/index.d.ts",
+  "main": "./dist/web-sdk.js",
+  "types": "./dist/src/web-sdk.d.ts",
   "scripts": {
-    "start": "webpack-dev-server --watch -d --env development",
+    "start": "webpack-dev-server --watch -d --env development src/web-sdk.tsx",
     "nonlocal": "npm run start -- --host 0.0.0.0",
+    "docs": "typedoc",
     "prepack": "npm run build",
-    "build": "rm -rf dist && mkdir dist && webpack -p --mode production --env production",
+    "build": "bash bin/build.sh",
     "analyze": "webpack -p --mode production --env analyze",
     "postversion": "bash bin/postversion.sh",
     "postpublish": "bash bin/deploy.sh",
@@ -53,6 +54,7 @@
     "gh-pages": "^2.1.1",
     "prettier": "^1.19.1",
     "ts-loader": "^6.2.1",
+    "typedoc": "^0.16.11",
     "typescript": "^3.7.3",
     "webpack": "^4.41.2",
     "webpack-bundle-analyzer": "^3.6.0",
