+++ src/TagListOverlay.js
@@ -1,4 +1,10 @@
 /** @jsx h */
+
+/**
+ * @internal
+ * @packageDocumentation
+ */
+
 // TODO (2018-09-17) Brian Mock
 // - Probably share some code with FloorOverlay eventually
