--- src/types.d.ts
@@ -1,3 +1,0 @@
-declare module "socket.io-client";
-
-declare const GLOBAL_VERSION: string;
