+++ typedoc.json
@@ -1,0 +1,9 @@
+{
+  "mode": "modules",
+  "inputFiles": ["src/web-sdk.tsx"],
+  "out": "./docs",
+  "readme": "docs-readme.md",
+  "stripInternal": true,
+  "excludeNotExported": true,
+  "excludePrivate": true
+}
